pyproject.toml
[tool.poetry]
name = "github-cli"
version = "0.1.0"
description = "Github Cli Using Typer"
authors = ["MahmudulHassan5809 <mahmudul.hassan240@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.12.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
black = "^24.4.2"
pytest = "^8.2.2"
mypy = "^1.10.0"
pylint = "^3.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
